Revert "Require doubles for width, height, x and y for more precision."
This reverts commit 9ae447525c050e90f16a846fb300a7e44a1fa31a.

+++ packages/native_pdf_renderer/lib/src/document.dart
@@ -94,8 +94,8 @@ class PDFDocument {
       document: this,
       id: obj['id'] as String,
       pageNumber: pageNumber,
-      width: obj['width'] as double,
-      height: obj['height'] as double,
+      width: obj['width'] as int,
+      height: obj['height'] as int,
     );
   }
